fix(stt): wire the language list into the picker that's actually live
A subagent review caught that the first commit only widened
SourceTranscriptModal, which NewEditorShell never mounts (LeftPanel is
always rendered with active="chat"). The picker a real user opens is
MediaStage.tsx's own, separate <select>, still hardcoded to auto/en/fr/es.

- Extract language-label resolution into
  lib/ai-edition/transcription/languageLabels.ts (languageLabel,
  sortedLanguageOptions) so both pickers share one implementation instead
  of drifting the way the original hand-duplicated lists did.
- Wire MediaStage.tsx's picker onto it.
- Fix the "detected language" pill in both components to show a localized
  name instead of the raw whisper code.
- Guard Intl.Collator the same way Intl.DisplayNames already was, and cache
  a failed Intl.DisplayNames construction instead of retrying it on every
  language in the list.
- Move TRANSCRIPT_LANGUAGE_NAMES out of the schema module (bundled into the
  Electron main process) into the new UI-facing module, and collapse three
  copies of `Exclude<TranscriptLanguageCode, "auto">` into one exported
  WhisperLanguageCode type.
