+++ src/lib/ai-edition/schema/index.ts
@@ -816,8 +816,14 @@ export const chatInputSchema = z.object({
 // (`src/whisper.cpp`, verified against the tag `nix/whisper-stt.nix` pins —
 // v1.9.1) — `wparams.language` in
 // electron/native/whisper-stt/src/main.cpp forwards the code verbatim, so a
-// value outside this list fails to resolve a language id there. The UI's
-// "Regenerate as" picker (Modals.tsx) is the only current consumer.
+// value outside this list fails to resolve a language id there. "auto" is
+// always index 0 — code that needs "every real language, no sentinel" relies
+// on that (see `WhisperLanguageCode` below) rather than filtering it out.
+// Consumed by both "Regenerate as" pickers: `MediaStage.tsx` (the one
+// actually mounted by `NewEditorShell`) and `Modals.tsx`'s
+// `SourceTranscriptModal` (currently unreachable — `LeftPanel` is only ever
+// mounted with `active="chat"` — but kept correct rather than deleted, since
+// nothing marks it dead code and a test or a future rewire could reach it).
 export const TRANSCRIPT_LANGUAGE_CODES = [
 	"auto",
 	"en",
@@ -924,117 +930,6 @@ export const TRANSCRIPT_LANGUAGE_CODES = [
 
 export const transcriptLanguageSchema = z.enum(TRANSCRIPT_LANGUAGE_CODES);
 
-/**
- * English fallback name per code, from whisper.cpp's own `g_lang` table.
- * The UI prefers `Intl.DisplayNames` in the active locale and falls back to
- * this when that can't resolve a name (e.g. an unusual code on an older ICU).
- */
-export const TRANSCRIPT_LANGUAGE_NAMES: Record<
-	Exclude<(typeof TRANSCRIPT_LANGUAGE_CODES)[number], "auto">,
-	string
-> = {
-	en: "English",
-	zh: "Chinese",
-	de: "German",
-	es: "Spanish",
-	ru: "Russian",
-	ko: "Korean",
-	fr: "French",
-	ja: "Japanese",
-	pt: "Portuguese",
-	tr: "Turkish",
-	pl: "Polish",
-	ca: "Catalan",
-	nl: "Dutch",
-	ar: "Arabic",
-	sv: "Swedish",
-	it: "Italian",
-	id: "Indonesian",
-	hi: "Hindi",
-	fi: "Finnish",
-	vi: "Vietnamese",
-	he: "Hebrew",
-	uk: "Ukrainian",
-	el: "Greek",
-	ms: "Malay",
-	cs: "Czech",
-	ro: "Romanian",
-	da: "Danish",
-	hu: "Hungarian",
-	ta: "Tamil",
-	no: "Norwegian",
-	th: "Thai",
-	ur: "Urdu",
-	hr: "Croatian",
-	bg: "Bulgarian",
-	lt: "Lithuanian",
-	la: "Latin",
-	mi: "Maori",
-	ml: "Malayalam",
-	cy: "Welsh",
-	sk: "Slovak",
-	te: "Telugu",
-	fa: "Persian",
-	lv: "Latvian",
-	bn: "Bengali",
-	sr: "Serbian",
-	az: "Azerbaijani",
-	sl: "Slovenian",
-	kn: "Kannada",
-	et: "Estonian",
-	mk: "Macedonian",
-	br: "Breton",
-	eu: "Basque",
-	is: "Icelandic",
-	hy: "Armenian",
-	ne: "Nepali",
-	mn: "Mongolian",
-	bs: "Bosnian",
-	kk: "Kazakh",
-	sq: "Albanian",
-	sw: "Swahili",
-	gl: "Galician",
-	mr: "Marathi",
-	pa: "Punjabi",
-	si: "Sinhala",
-	km: "Khmer",
-	sn: "Shona",
-	yo: "Yoruba",
-	so: "Somali",
-	af: "Afrikaans",
-	oc: "Occitan",
-	ka: "Georgian",
-	be: "Belarusian",
-	tg: "Tajik",
-	sd: "Sindhi",
-	gu: "Gujarati",
-	am: "Amharic",
-	yi: "Yiddish",
-	lo: "Lao",
-	uz: "Uzbek",
-	fo: "Faroese",
-	ht: "Haitian Creole",
-	ps: "Pashto",
-	tk: "Turkmen",
-	nn: "Norwegian Nynorsk",
-	mt: "Maltese",
-	sa: "Sanskrit",
-	lb: "Luxembourgish",
-	my: "Myanmar",
-	bo: "Tibetan",
-	tl: "Tagalog",
-	mg: "Malagasy",
-	as: "Assamese",
-	tt: "Tatar",
-	haw: "Hawaiian",
-	ln: "Lingala",
-	ha: "Hausa",
-	ba: "Bashkir",
-	jw: "Javanese",
-	su: "Sundanese",
-	yue: "Cantonese",
-};
-
 export type AxcutWord = z.infer<typeof wordSchema>;
 export type AxcutTranscriptSegment = z.infer<typeof transcriptSegmentSchema>;
 export type AxcutTranscript = z.infer<typeof transcriptSchema>;
@@ -1056,6 +951,8 @@ export type CreateProjectInput = z.infer<typeof createProjectInputSchema>;
 export type AddAssetInput = z.infer<typeof addAssetInputSchema>;
 export type ChatInput = z.infer<typeof chatInputSchema>;
 export type TranscriptLanguageCode = z.infer<typeof transcriptLanguageSchema>;
+/** A real whisper.cpp language code — `TranscriptLanguageCode` minus the "auto" detection sentinel. */
+export type WhisperLanguageCode = Exclude<TranscriptLanguageCode, "auto">;
 
 export function createEmptyDocument(
 	input: CreateProjectInput & { projectId: string; createdAt?: string },
